test: complete

+++ routes/api.js
@@ -24,9 +24,9 @@ module.exports = function (app) {
     .get(async function (req, res){
       let project = req.params.project;
 
-      const allIssues = await Issue.find({})
+      const foundIssues = await Issue.find(req.query)
 
-      res.send(allIssues)
+      res.send(foundIssues)
     })
     
     .post(async function (req, res){
@@ -55,13 +55,32 @@ module.exports = function (app) {
       let project = req.params.project;
       const {_id, issue_title, issue_text, created_by, assigned_to, status_text, open} = req.body
 
+      if(!_id) {
+        return res.send('Please provide an ID...')
+      }
+
+      function isAllFieldsButIdEmpty(obj) {
+        for (let key in obj) {
+          if (key !== '_id' && obj.hasOwnProperty(key)) {
+            if (obj[key].length !== 0) {
+              return false
+            }
+          }
+        }
+        return true
+      }
+
+      if (isAllFieldsButIdEmpty(req.body)) {
+        return res.send('Please, provide a field to update...')
+      }
+
       const updatedIssue = {
         ...(issue_title && {issue_title}),
         ...(issue_text && {issue_text}),
         ...(created_by && {created_by}),
         ...(assigned_to && {assigned_to}),
         ...(status_text && {status_text}),
-        ...(open && {open: false}),
+        ...(open && {open}),
         updated_on: new Date().toISOString()
       }
 
@@ -69,7 +88,6 @@ module.exports = function (app) {
         await Issue.findByIdAndUpdate(_id, updatedIssue)
         return res.send('Updated in database!')
       } catch(err) {
-        console.log(err)
         return res.send('No match found in database :(')
       }
     })
@@ -77,14 +95,23 @@ module.exports = function (app) {
     .delete(async function (req, res){
       let project = req.params.project;
       const {_id} = req.body
+
+      if (!_id) {
+        return res.send('Please provide an ID')
+      }
       
       try {
         await Issue.findByIdAndDelete(_id)
         return res.send('Removed from database!')
       } catch(err) {
-        console.log(err)
+        // console.log(err)
         return res.send('No match found in database :(')
       }
     });
     
+    app.delete('/api/deleteAll', async (req, res) => {
+      const deletedCount = await Issue.deleteMany({})
+      res.send(`deleted all documents from database + ${JSON.stringify(deletedCount)}`)
+    })
+
 };
